+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ability to view sublattices list as part of the main lattice
 - Ability to view subalattices graph as part of main lattice
 
+
+### Fixed
+
+- Electron dependencies are no longer written twice to the DB during a workflow
+
 ## [0.201.0] - 2022-10-09
 
 ### Authors